Move sdp.h to protocols directory
Moved sdp.h to protocols directory and make necessary changes
to support this move.
